refactor: ProfileDto를 Profile로 수정
