+++ src/main/java/real/world/domain/profile/query/ProfileQueryRepository.java
@@ -1,0 +1,9 @@
+package real.world.domain.profile.query;
+
+import java.util.Optional;
+
+public interface ProfileQueryRepository {
+
+    Optional<Profile> findByLoginIdAndUsername(Long loginId, String username);
+
+}
