+++ src/main/java/real/world/domain/profile/query/ProfileQueryRepositoryImpl.java
@@ -1,4 +1,4 @@
-package real.world.domain.profile.repository;
+package real.world.domain.profile.query;
 
 import static real.world.domain.follow.entity.QFollow.follow;
 import static real.world.domain.user.entity.QUser.user;
@@ -7,7 +7,6 @@ import com.querydsl.jpa.JPAExpressions;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import java.util.Optional;
 import org.springframework.stereotype.Repository;
-import real.world.domain.user.dto.ProfileDto;
 import real.world.domain.user.dto.QProfileDto;
 
 @Repository
@@ -20,7 +19,7 @@ public class ProfileQueryRepositoryImpl implements ProfileQueryRepository {
     }
 
     @Override
-    public Optional<ProfileDto> findByLoginIdAndUsername(Long loginId, String username) {
+    public Optional<Profile> findByLoginIdAndUsername(Long loginId, String username) {
         return Optional.ofNullable(
             factory
                 .select(
