--- src/main/java/real/world/domain/profile/repository/ProfileQueryRepository.java
@@ -1,10 +1,0 @@
-package real.world.domain.profile.repository;
-
-import java.util.Optional;
-import real.world.domain.user.dto.ProfileDto;
-
-public interface ProfileQueryRepository {
-
-    Optional<ProfileDto> findByLoginIdAndUsername(Long loginId, String username);
-
-}
