+++ src/main/java/real/world/domain/profile/service/ProfileQueryService.java
@@ -2,8 +2,8 @@ package real.world.domain.profile.service;
 
 import org.springframework.stereotype.Service;
 import real.world.domain.profile.dto.response.ProfileResponse;
-import real.world.domain.profile.repository.ProfileQueryRepository;
-import real.world.domain.user.dto.ProfileDto;
+import real.world.domain.profile.query.ProfileQueryRepository;
+import real.world.domain.profile.query.Profile;
 import real.world.error.exception.UsernameNotExistException;
 
 @Service
@@ -16,7 +16,7 @@ public class ProfileQueryService {
     }
 
     public ProfileResponse getProfile(Long loginId, String username) {
-        final ProfileDto profile = profileQueryRepository.findByLoginIdAndUsername(loginId, username)
+        final Profile profile = profileQueryRepository.findByLoginIdAndUsername(loginId, username)
             .orElseThrow(UsernameNotExistException::new);
         return ProfileResponse.of(profile);
     }
